+++ CHANGELOG.md
@@ -1281,3 +1281,4 @@
 - [2026-05-18 21:23] feat(state): integrate custom telemetry logger in state
 - [2026-05-18 21:48] feat(network): integrate offline caching for network state
 - [2026-05-18 21:52] perf(forms): apply RepaintBoundary on animated forms widgets
+- [2026-05-18 21:57] fix(theme): fix memory leak by cancelling theme subscription